Fixed type

+++ WebService.API/Datas/Models/Products/Cpu.cs
@@ -8,10 +8,10 @@ public record Cpu : IProduct
 	public int Tdp { get; set; }
 	public int Core { get; set; }
 	public int Thread { get; set; }
-	public float Frequency { get; set; }
+	public double Frequency { get; set; }
 	public string? Manufacturer { get; set; }
-	public required long Id { get; set; }
+	public required int Id { get; set; }
 	public string? Name { get; set; }
 	public string? Description { get; set; }
-	public float BasePrice { get; set; }
+	public double BasePrice { get; set; }
 }
