+++ WebService.API/VirtualBase/IPromoteStrategy.cs
@@ -4,7 +4,7 @@ public interface IPromoteStrategy
 {
 	public string Details { get; }
 
-	public float DoDiscount(float basePrice);
+	public double DoDiscount(double basePrice);
 
 	public bool CheckCondition(IProduct product);
 }
